requirements.txt
# Competition infrastructure requirements
# ========================================
# These are the minimal dependencies needed by the CI pipeline
# (competition/evaluate.py, validate_submission.py, render_leaderboard.py,
#  encryption/encrypt.py, encryption/decrypt.py).
#
# For training GNN models, see starter_code/requirements.txt

pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pyyaml>=6.0
cryptography>=41.0.0
